[BUGFIX] wifi: cfg80211: free beacon_ies when overriden from hidden BSS
This is a more of a cosmetic fix. The branch will only be taken if
proberesp_ies is set, which implies that beacon_ies is not set unless we
are connected to an AP that just did a channel switch. And, in that case
we should have found the BSS in the internal storage to begin with.

type=bugfix
ticket=none
fixes=unknown

Change-Id: Id8c4c10aedd176ef2e18a4cad747b299f150f9df
Signed-off-by: Benjamin Berg <benjamin.berg@intel.com>
Reviewed-on: https://gerritwcs.ir.intel.com/c/iwlwifi-stack-dev/+/78367
tested: iil_jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-by: Berg, Johannes <johannes.berg@intel.com>
x-iwlwifi-stack-dev: fceabc93a09c5b5ead41b3921f0819324a68f9c9