[BUGFIX] wifi: iwlwifi: fix firmware dump at device stop
Due to the now asynchronous nature of the firmware dump and
locking involved in the opmode, it's possible that the worker
is scheduled just as the device is stopping, and the opmode
has already acquired the locks. This then blocks the worker,
which will only try to dump the firmware info after it was
already stopped, obviously failing.

Fix this by passing the struct to the dump callback, and in
the struct (which is a pointer so it can be changed after
the function is already started and blocked on the mutex)
indicate the context in which the function was called. This
also gets rid of the special case for reset handshake, but
more importantly lets us fix the problem by making a call
to do the dump if it was pending, and aborting the one from
the worker by setting the context to abort. That requires a
bit more work from the opmode, but doesn't seem too bad.

type=bugfix
fixes=I96a9243e7877808ed6d1bff6967c15d6c24882f0
ticket=none

Change-Id: Ia2826544bc3e1de1421ad75e08a26a6aefb057a6
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://gerritwcs.ir.intel.com/c/iwlwifi-stack-dev/+/154452
automatic-review: iil_jenkins iil_jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
tested: iil_jenkins iil_jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Tested-by: iil_jenkins iil_jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-by: Miriam Rachel Korenblit <miriam.rachel.korenblit@intel.com>
Reviewed-on: https://gerritwcs.ir.intel.com/c/iwlwifi-stack-dev/+/159053
x-iwlwifi-stack-dev: babfa53a586b27b209a9a5869fcffad0d82c9acc