[BUGFIX] wifi: iwlwifi: don't warn during reprobe
During reprobe, the sw state is being destroyd, and so is the
connection. When the peer STA is being removed, the opmode sends a
command to flush the TXQs of the STA and uses iwl_trans_wait_txq_empty.

This one warns if the FW is not alive, but it really shouldn't if
there is a FW error - and return silently instead, just like we do when
sending a hcmd.

type=bugfix
ticket=none
fixes=unknown

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Change-Id: I3bf0de2eb090a8b94c4e36d93dd91df61fadb808
Reviewed-on: https://gerritwcs.ir.intel.com/c/iwlwifi-stack-dev/+/156927
tested: iil_jenkins iil_jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://gerritwcs.ir.intel.com/c/iwlwifi-stack-dev/+/159054
x-iwlwifi-stack-dev: 87375598a78c6181b8acff8bf7d73fe5c5a6cdf2