backport: improve debugfs cancellation backport
This was backported as a no-op, which in wireless results in
simply waiting for the worker, but never being able to cancel
it. This is fine functionally, but leaves the deadlock that
this infrastructure was meant to prevent. A better backport
is possible: add the cancellations to a list and cancel if
needed when a file is being removed (recursively).

Of course this cannot be completely race free as a new open
could happen while the removal is in progress, but it should
fix the (already rare) deadlocks in most cases.

type=feature
ticket=none

Change-Id: I02ba2e5ac6582584d8a94c9f200889ff65ce4ccd
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://gerritwcs.ir.intel.com/c/iwlwifi-stack-dev/+/174710
tested: iil_jenkins iil_jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Tested-by: iil_jenkins iil_jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-by: Miriam Rachel Korenblit <miriam.rachel.korenblit@intel.com>
Reviewed-on: https://gerritwcs.ir.intel.com/c/iwlwifi-stack-dev/+/178212
x-iwlwifi-stack-dev: bf210fcd46b8d04858a9537387a7d1034fead3c8