wifi: iwlwifi: cfg: move MAC parameters to base MAC data
There are a number of MAC parameters that are in the iwl_cfg
(which is the last config matched to the MAC/RF combination).
This isn't necessary, there are many more of those than MACs,
so move (most of) the data into the MAC base config struct.

Note that DCCM information remains for use by older devices,
and on 9000 series it'll be in struct iwl_cfg but be ignored
when the CRF is in a Qu/So platform.

type=cleanup
ticket=none

Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Change-Id: I68e5d9d640a1827549c3052f66c441c606bb8d68
x-iwlwifi-stack-dev: b276abb4a0a5ad983a7ff4e562df169ddb37e5e1