wifi: iwlwifi: rename struct iwl_base_params to iwl_mac_base
These are (going to be) base MAC parameters that are identical
even for different platforms with the same MAC, so rename the
structure accordingly.

Also rename the pointer to it so the dereferencing is a bit
shorter.

type=cleanup
ticket=none

Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Change-Id: I1108eb51a4935abec4206f0aac461a3833033e79
x-iwlwifi-stack-dev: 6762e25ae022b0ea2cff2daee2bda94873dfa507