wifi: iwlwifi: cfg: remove MAC type/step matching
Now that it's all split into MAC and RF configs, remove
the matching on MAC type and step. If we ever need to do
something based on the MAC step, we'll have to find some
new mechanism (since the MAC type is known already from
the PCI IDs table, but not the step), or just handle the
(likely small) differences in code.

We could possibly remove cores/jacket bits, but we can't
make the struct smaller, so leave them in case they may
be useful in the future.

type=cleanup
ticket=none

Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Change-Id: Iae27b781221fd29845493adf2c29d9e4f7a9c33b
x-iwlwifi-stack-dev: 676202244acfa26042e0a9f527dc50a7a52609d8