[BUGFIX] wifi: iwlwifi: mvm: avoid outdated reorder buffer head_sn
If no frames are received on a queue for a while, the reorder buffer
head_sn may be an old one. When the next frame that is received on
that queue and buffered is a subframe of an AMSDU but not the last
subframe, it will not update the buffer's head_sn. When the frame
release notification arrives, it will not release the buffered frame
because it will look like the notification's NSSN is lower than the
buffer's head_sn (because of a wraparound).
Fix it by updating the head_sn when the first frame is buffered.

type=bugfix
ticket=jira:WIFI-466778
fixes=I2c5b01c0655914d41cc2fc2eb11387b5e7134db6

Signed-off-by: Avraham Stern <avraham.stern@intel.com>
Change-Id: I9ec9a3508e7935e8d1833ea3e086066fdefee644
Reviewed-on: https://gerritwcs.ir.intel.com/c/iwlwifi-stack-dev/+/204896
Reviewed-by: Daniel Gabay <daniel.gabay@intel.com>
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-by: Miriam Rachel Korenblit <miriam.rachel.korenblit@intel.com>
tested: iil_jenkins iil_jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Tested-by: iil_jenkins iil_jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
x-iwlwifi-stack-dev: 827a24719d6b3e0d958ffcfb6d83be384a1674c5