[BUGFIX] wifi: iwlwifi: mld: avoid outdated reorder buffer head_sn
If no frames are received on a queue for a while, the reorder buffer
head_sn may be an old one. When the next frame that is received on
that queue and buffered is a subframe of an AMSDU but not the last
subframe, it will not update the buffer's head_sn. When the frame
release notification arrives, it will not release the buffered frame
because it will look like the notification's NSSN is lower than the
buffer's head_sn (because of a wraparound).
Fix it by updating the head_sn when the first frame is buffered.

type=bugfix
ticket=jira:WIFI-466778
fixes=unknown

Signed-off-by: Avraham Stern <avraham.stern@intel.com>
Change-Id: I7b57a481122074b1f40d39cd31db2e5262668eb2
Reviewed-on: https://gerritwcs.ir.intel.com/c/iwlwifi-stack-dev/+/205161
tested: iil_jenkins iil_jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Tested-by: iil_jenkins iil_jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-by: Miriam Rachel Korenblit <miriam.rachel.korenblit@intel.com>
x-iwlwifi-stack-dev: d1a4221194f8b5b5b9038887737d9ea11708775c