[BUGFIX] wifi: iwlwifi: pcie: allow not unmapping pages
For devices that are somehow marked untrusted (external devices, or
explicitly via the BIOS), the unmap of the DMA will cause the IOMMU
to flush the mapping, which requires locking and waiting for the
IOMMU hardware. In order to work around that, we want to (in that
case only) copy the data into the SKB and not use page/frag RX.

As a first step, add a transport configuration to be able to opt in
by the opmode (only iwlmld will, for now) and set that the RX data
should not be automatically unmapped. The opmode will need to be
much more careful, but the mapping can be reused if the page wasn't
used for the SKB, avoiding the high IOMMU cost later.

type=bugfix
ticket=jira:WIFI-682219
ticket=jira:WIFI-643529
fixes=unknown

Change-Id: I11d9e2e926db07f3518003a43a4309467d262127
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://gerritwcs.ir.intel.com/c/iwlwifi-stack-dev/+/211966
tested: iil_jenkins iil_jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-by: Benjamin Berg <benjamin.berg@intel.com>
Reviewed-by: Miriam Rachel Korenblit <miriam.rachel.korenblit@intel.com>
x-iwlwifi-stack-dev: 50848855bc3bd7b9720ae35fa103a541aa1b4d48