wifi: iwlwifi: pcie: move pm_support to the specific transport
Currently it is under iwl_trans, which is the bus agnostic part of the
transport. But really it is relevant for pcie only, so move it to the
iwl_trans_pcie and export it via an API to the opmode.

type=feature
ticket=none

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Change-Id: Ic6648f040430c94150d0fa11601f50a6a630b862
Reviewed-on: https://gerritwcs.ir.intel.com/c/iwlwifi-stack-dev/+/216551
tested: iil_jenkins iil_jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
x-iwlwifi-stack-dev: cfdfdbc8994a2f7dba1ee3e641175981b421d1a7