wifi: iwlwifi: rename iwl_finish_nic_init
The function is called so because for older devices it sets a bit called
"init_done". But for the latest devices it sets a different bit,
"mac_init". Since this name is not clear anyway, rename it such that it
indicates the logic of the newer devices.
Also add the 'trans' prefix so iw will be clear from the name that this
is a transport API.

type=feature
ticket=none

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Change-Id: Ibd41b0c8f7fbae77026e76dbbc085df3eecec538
Reviewed-on: https://gerritwcs.ir.intel.com/c/iwlwifi-stack-dev/+/216549
tested: iil_jenkins iil_jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
x-iwlwifi-stack-dev: b15ec6d3cd49173c91fa3684b931aaf8c04b4c9c